refactor(daemon): decompose init() into phase methods in partition files
init() becomes a byte-order-faithful phase sequence; start/stop and the
bridge warnings move to daemon/lifecycle.cpp.
